backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
PyMuPDF==1.24.0
langchain==0.2.14
langchain-google-genai==1.0.8
langchain-community==0.2.12
langchain-text-splitters==0.2.2
sentence-transformers==3.0.0
faiss-cpu==1.8.0
python-dotenv==1.0.1
